Move `EventStore` under sub-package
