+++ server/src/main/java/io/spine/server/event/EventBus.java
@@ -40,6 +40,7 @@ import io.spine.server.bus.BusBuilder;
 import io.spine.server.bus.DeadMessageHandler;
 import io.spine.server.bus.EnvelopeValidator;
 import io.spine.server.bus.MulticastBus;
+import io.spine.server.event.store.EventStore;
 import io.spine.server.storage.StorageFactory;
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
