+++ server/src/test/java/io/spine/server/event/given/EventStoreTestEnv.java
@@ -26,8 +26,8 @@ import com.google.protobuf.Timestamp;
 import io.grpc.stub.StreamObserver;
 import io.spine.core.Event;
 import io.spine.server.BoundedContext;
-import io.spine.server.event.EventStore;
-import io.spine.server.event.EventStoreTest;
+import io.spine.server.event.store.EventStore;
+import io.spine.server.event.store.EventStoreTest;
 import io.spine.test.event.ProjectCreated;
 import io.spine.test.event.TaskAdded;
 import io.spine.testdata.Sample;
