+++ server/src/test/java/io/spine/server/event/store/ERepositoryTest.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.event;
+package io.spine.server.event.store;
 
 import com.google.protobuf.Any;
 import com.google.protobuf.util.Timestamps;
@@ -26,13 +26,15 @@ import io.spine.client.ColumnFilter;
 import io.spine.client.CompositeColumnFilter;
 import io.spine.client.CompositeColumnFilter.CompositeOperator;
 import io.spine.client.EntityFilters;
+import io.spine.server.event.EventFilter;
+import io.spine.server.event.EventStreamQuery;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
 import static io.spine.protobuf.TypeConverter.toObject;
-import static io.spine.server.event.ERepository.toEntityFilters;
+import static io.spine.server.event.store.ERepository.toEntityFilters;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
