+++ server/src/test/java/io/spine/server/event/store/MatchFilterTest.java
@@ -18,8 +18,9 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.event;
+package io.spine.server.event.store;
 
+import io.spine.server.event.EventFilter;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -28,7 +29,6 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 @DisplayName("MatchFilter should")
 class MatchFilterTest {
 
-    @SuppressWarnings("ConstantConditions") // Passing `null` is the purpose of the test.
     @Test
     @DisplayName("not allow null events to pass")
     void notAllowNullEvents() {
